RText!

+++ QuickBackupM_en.py
@@ -1,15 +1,14 @@
 # coding: utf8
-import copy
-import json
 import os
+import re
 import shutil
-import sys
 import time
 from threading import Lock
+from utils.rtext import *
 
 
 '''================ Modifiable constant starts ================'''
-sizeDisplay = True
+SizeDisplay = True
 SlotCount = 5
 Prefix = '!!qb'
 BackupPath = './qb_multi'
@@ -32,23 +31,22 @@ OverwriteBackupFolder = 'overwrite'
 ServerPath = './server'
 '''================ Modifiable constant ends ================'''
 
-
 HelpMessage = '''
 ------MCDR Multi Quick Backup------
 A plugin that supports multi slots world §abackup§r and backup §crestore§r
 §a[Format]§r
 §7{0}§r Display help message
-§7{0} make §e[<comment>]§r Make a §abackup§r to slot §61§r, and shift the slots behind。§e<comment>§r is an optional comment message
-§7{0} back §6[<slot>]§r §cRestore§r the world to slot §61§r. When §6<slot>§r parameter is set it will §crestore§r to slot §6<slot>§r
+§7{0} make §e[<comment>]§r Make a §abackup§r to slot §61§r. §e<comment>§r is an optional comment message
+§7{0} back §6[<slot>]§r §cRestore§r the world to slot §6<slot>§r
 §7{0} del §6[<slot>]§r §cDelete§r the world in slot §6<slot>§r
 §7{0} confirm§r Use after execute back to confirm §crestore§r execution
 §7{0} abort§r Abort backup §crestoring§r
 §7{0} list§r Display slot information
 When §6<slot>§r is not set the default value is §61§r
 §a[Example]§r
-§7{0} make
+§7{0} make§r
 §7{0} make §eworld eater done§r
-§7{0} back
+§7{0} back§r
 §7{0} back §62§r
 '''.strip().format(Prefix)
 slot_selected = None
@@ -58,7 +56,7 @@ plugin_unloaded = False
 creating_backup = Lock()
 restoring_backup = Lock()
 '''
-MCDR_root/
+mcdr_root/
 	server/
 		world/
 	qb_multi/
@@ -74,15 +72,16 @@ MCDR_root/
 '''
 
 
-def print_message(server, info, msg, tell=True):
-	for line in msg.splitlines():
-		if info.isPlayer:
-			if tell:
-				server.tell(info.player, line)
-			else:
-				server.say(line)
-		else:
-			print(line)
+def print_message(server, info, msg, tell=True, prefix='[QBM] '):
+	msg = prefix + msg
+	if info.is_player and not tell:
+		server.say(msg)
+	else:
+		server.reply(info, msg)
+
+
+def command_run(message, text, command):
+	return RText(message).set_hover_text(text).set_click_event(RAction.run_command, command)
 
 
 def copy_worlds(src, dst):
@@ -97,11 +96,6 @@ def remove_worlds(folder):
 		shutil.rmtree('{}/{}'.format(folder, world))
 
 
-def info_message(server, info, msg, tell=False):
-	for line in msg.splitlines():
-		print_message(server, info, '[QBM] ' + line, tell)
-
-
 def get_slot_folder(slot):
 	return '{}/slot{}'.format(BackupPath, slot)
 
@@ -112,8 +106,6 @@ def get_slot_info(slot):
 			info = json.load(f, encoding='utf8')
 		for key in info.keys():
 			value = info[key]
-			if sys.version_info.major == 2 and type(value) is unicode:
-				info[key] = value.encode('utf8')
 	except:
 		info = None
 	return info
@@ -147,26 +139,53 @@ def touch_backup_folder():
 		mkdir(get_slot_folder(i + 1))
 
 
+def slot_number_formater(slot):
+	flag_fail = False
+	if type(slot) is not int:
+		try:
+			slot = int(slot)
+		except ValueError:
+			flag_fail = True
+	if flag_fail or not 1 <= slot <= SlotCount:
+		return None
+	return slot
+
+
+def slot_check(server, info, slot):
+	slot = slot_number_formater(slot)
+	if slot is None:
+		print_message(server, info, 'Slot format wrong, it should be a number between [{}, {}]'.format(1, SlotCount))
+		return None
+
+	slot_info = get_slot_info(slot)
+	if slot_info is None:
+		print_message(server, info, 'Slot §6{}§r is empty'.format(slot))
+		return None
+	return slot, slot_info
+
+
 def delete_backup(server, info, slot):
 	global creating_backup, restoring_backup
 	if creating_backup.locked() or restoring_backup.locked():
 		return
+	if slot_check(server, info, slot) is None:
+		return
 	try:
 		shutil.rmtree(get_slot_folder(slot))
-	except:
-		info_message(server, info, "§4Delete fail, check console for more detail")
+	except Exception as e:
+		print_message(server, info, RText('§Delete fail§r, check console for more detail').set_hover_text(e))
 	else:
-		info_message(server, info, "§aDelete success")
+		print_message(server, info, '§aDelete success§r')
 
 
 def create_backup(server, info, comment):
 	global creating_backup
 	acquired = creating_backup.acquire(blocking=False)
 	if not acquired:
-		info_message(server, info, '§aBacking up§r, dont spam')
+		print_message(server, info, '§aBacking up§r, don''t spam')
 		return
 	try:
-		info_message(server, info, '§aBacking up§r, please wait')
+		print_message(server, info, '§aBacking up§r, please wait')
 		start_time = time.time()
 		touch_backup_folder()
 
@@ -197,44 +216,16 @@ def create_backup(server, info, comment):
 		if comment is not None:
 			slot_info['comment'] = comment
 		with open('{}/info.json'.format(slot_path), 'w') as f:
-			if sys.version_info.major == 2:
-				json.dump(slot_info, f, indent=4, encoding='utf8')
-			else:
-				json.dump(slot_info, f, indent=4)
+			json.dump(slot_info, f, indent=4)
 		end_time = time.time()
-		info_message(server, info, '§aBack up§r successfully, time cost ' + str(end_time - start_time)[:3] + 's')
-		info_message(server, info, format_slot_info(info_dict=slot_info))
+		print_message(server, info, '§aBack up§r successfully, time cost §6{}§rs'.format(round(end_time - start_time, 1)))
+		print_message(server, info, format_slot_info(info_dict=slot_info))
 	except Exception as e:
-		info_message(server, info, '§aBack up§r unsuccessfully, error code {}'.format(e))
+		print_message(server, info, '§aBack up§r unsuccessfully, error code {}'.format(e))
 	finally:
 		creating_backup.release()
 		if TurnOffAutoSave:
 			server.execute('save-on')
-
-
-def slot_number_formater(slot):
-	flag_fail = False
-	if type(slot) is not int:
-		try:
-			slot = int(slot)
-		except ValueError:
-			flag_fail = True
-	if flag_fail or not 1 <= slot <= SlotCount:
-		return None
-	return slot
-
-
-def slot_check(server, info, slot):
-	slot = slot_number_formater(slot)
-	if slot is None:
-		info_message(server, info, 'Slot format wrong, it should be a number between [{}, {}]'.format(1, SlotCount))
-		return None
-
-	slot_info = get_slot_info(slot)
-	if slot_info is None:
-		info_message(server, info, 'Slot is empty')
-		return None
-	return slot, slot_info
 
 
 def restore_backup(server, info, slot):
@@ -246,156 +237,173 @@ def restore_backup(server, info, slot):
 	global slot_selected, abort_restore
 	slot_selected = slot
 	abort_restore = False
-	info_message(server, info, 'Gonna restore the world to slot §6{}§r, {}'.format(slot, format_slot_info(info_dict=slot_info)))
-	info_message(server, info, 'Use §7{0} confirm§r to confirm §crestore§r, §7{0} abort§r to abort'.format(Prefix))
+	print_message(server, info, 'Gonna restore the world to slot §6{}§r, {}'.format(slot, format_slot_info(info_dict=slot_info)))
+	print_message(
+		server, info,
+		command_run('Use §7{0} confirm§r to confirm §crestore§r'.format(Prefix), 'click to confirm', '{0} confirm'.format(Prefix))
+		+ ', '
+		+ command_run('§7{0} abort§r to abort'.format(Prefix), 'click to abort', '{0} abort'.format(Prefix))
+	)
 
 
 def confirm_restore(server, info):
 	global restoring_backup
 	acquired = restoring_backup.acquire(blocking=False)
 	if not acquired:
-		info_message(server, info, '§cRestoring§r, dont spam')
+		print_message(server, info, '§cRestoring§r, don''t spam')
 		return
 	try:
 		global slot_selected
 		if slot_selected is None:
-			info_message(server, info, 'Nothing to confirm')
+			print_message(server, info, 'Nothing to confirm')
 			return
 		slot = slot_selected
 		slot_selected = None
 
-		info_message(server, info, '§cRestore§r after 10 second')
+		print_message(server, info, '§cRestore§r after 10 second')
 		for countdown in range(1, 10):
-			info_message(server, info, '{} second later the world will be §crestored§r to slot §6{}§r, {}'.format(10 - countdown, slot, format_slot_info(slot_number=slot)))
+			print_message(server, info, command_run(
+				'{} second later the world will be §crestored§r to slot §6{}§r, {}'.format(10 - countdown, slot, format_slot_info(slot_number=slot)),
+				'click to abort restore!',
+				'{} abort'.format(Prefix)
+			))
 			for i in range(10):
 				time.sleep(0.1)
 				global abort_restore
 				if abort_restore:
-					info_message(server, info, '§cRestore§r aborted!')
+					print_message(server, info, '§cRestore§r aborted!')
 					return
 
-		kick_bots(server, info)
 		server.stop()
-		# MCDaemon
-		if sys.version_info.major == 2:
-			print('[QBM] Wait for up to 10s for server to stop')
-			time.sleep(10)
-		# MCDReforged
-		else:
-			print('[QBM] Wait for server to stop')
-			while server.is_server_running():
-				time.sleep(0.1)
+		server.logger.info('[QBM] Wait for server to stop')
+		server.wait_for_start()
 
-		print('[QBM] Backup current world to avoid idiot')
+		server.logger.info('[QBM] Backup current world to avoid idiot')
 		overwrite_backup_path = BackupPath + '/' + OverwriteBackupFolder
 		if os.path.exists(overwrite_backup_path):
 			shutil.rmtree(overwrite_backup_path)
 		copy_worlds(ServerPath, overwrite_backup_path)
 		with open('{}/info.txt'.format(overwrite_backup_path), 'w') as f:
-			f.write('Overwrite time: {}\nConfirm by: {}'.format(format_time(),
-																info.player if info.isPlayer else '$Console$'))
+			f.write('Overwrite time: {}\n'.format(format_time()))
+			f.write('Confirmed by: {}'.format(info.player if info.is_player else '$Console$'))
 
 		slot_folder = get_slot_folder(slot)
-		print('[QBM] Restoring backup ' + slot_folder)
+		server.logger.info('[QBM] Deleting world')
 		remove_worlds(ServerPath)
-		time.sleep(0.5)
+		server.logger.info('[QBM] Restore backup ' + slot_folder)
 		copy_worlds(slot_folder, ServerPath)
-		print('[QBM] Wait for another 1s before server starts')
-		time.sleep(1)
 
 		server.start()
 	finally:
 		restoring_backup.release()
 
 
-def kick_bots(server, info):
-	try:
-		import mcdbot
-		import copy
-		iinfo = copy.deepcopy(info)
-		iinfo.content = '!!bot kickall'
-		mcdbot.onServerInfo(server, iinfo)
-	except:
-		pass
-
-def getdirsize(dir):
-	size = 0
-	for root, dirs, files in os.walk(dir):
-		size += sum([os.path.getsize(os.path.join(root, name)) for name in files])
-	return round((size/(1024*1024*1024)) , 4)
-	
 def list_backup(server, info):
-	global sizeDisplay
+	def get_dir_size(dir):
+		size = 0
+		for root, dirs, files in os.walk(dir):
+			size += sum([os.path.getsize(os.path.join(root, name)) for name in files])
+		if size < 2 ** 30:
+			return f'{round(size / 2 ** 20, 2)} MB'
+		else:
+			return f'{round(size / 2 ** 30, 2)} GB'
+
+	print_message(server, info, f'There are §6{SlotCount}§r slots')
 	for i in range(SlotCount):
-		info_message(server, info, '[Slot §6{}§r] {}'.format(i + 1, format_slot_info(slot_number=i + 1)))
-	if sizeDisplay:
-		size = getdirsize(BackupPath)
-		info_message(server, info, 'Space taken up by backup files：{}{} §rGB'.format(color, size))
+		j = i + 1
+		print_message(
+			server, info,
+			RTextList(
+				f'[Slot §6{j}§r] ',
+				RText('[▷] ', color=RColor.green)
+					.set_hover_text(f'click to restore to slot §6{j}§r')
+					.set_click_event(RAction.run_command, f'{Prefix} back {j}'),
+				RText('[×] ', color=RColor.red)
+					.set_hover_text(f'click to delete slot §6{j}§r')
+					.set_click_event(RAction.suggest_command, f'{Prefix} del {j}'),
+				format_slot_info(slot_number=j)
+			),
+			prefix=''
+		)
+	if SizeDisplay:
+		print_message(server, info, 'Total space consumed: §a{}§r'.format(get_dir_size(BackupPath)))
 
 
 def trigger_abort(server, info):
 	global abort_restore, slot_selected
 	abort_restore = True
 	slot_selected = None
-	info_message(server, info, 'Operation terminated!')
+	print_message(server, info, 'Operation terminated!')
 
 
-def onServerInfo(server, info):
-	content = info.content
-	if content == 'Saved the game' and not info.isPlayer:
-		global game_saved
-		game_saved = True
-	if not info.isPlayer and content.endswith('<--[HERE]'):
-		content = content.replace('<--[HERE]', '')
-
-	command = content.split()
-	if len(command) == 0 or command[0] != Prefix:
-		return
-	del command[0]
-
-	if len(command) == 0:
-		print_message(server, info, HelpMessage)
-		return
-
-	cmdLen = len(command)
-	# MCDR permission check
-	global MinimumPermissionLevel
-	if hasattr(server, 'MCDR') and cmdLen >= 1 and command[0] in MinimumPermissionLevel.keys():
-		if server.get_permission_level(info) < MinimumPermissionLevel[command[0]]:
-			print_message(server, info, '§cPermission denied§r')
-			return
-	# make [<comment>]
-	if cmdLen >= 1 and command[0] == 'make':
-		comment = content.replace('{} make'.format(Prefix), '', 1).lstrip(' ') if cmdLen >= 1 else None
-		create_backup(server, info, comment)
-	# back [<slot>]
-	elif cmdLen in [1, 2] and command[0] == 'back':
-		restore_backup(server, info, command[1] if cmdLen == 2 else '1')
-	# confirm
-	elif cmdLen == 1 and command[0] == 'confirm':
-		confirm_restore(server, info)
-	# abort
-	elif cmdLen == 1 and command[0] == 'abort':
-		trigger_abort(server, info)
-	# list
-	elif cmdLen == 1 and command[0] == 'list':
-		list_backup(server, info)
-	# delete
-	elif cmdLen in [1, 2] and command[0] == 'del':
-		delete_backup(server, info, command[1] if cmdLen == 2 else '1')
-
-	else:
-		print_message(server, info, 'Unknown command, input §7' + Prefix + '§r for help')
+def print_help_message(server, info):
+	for line in HelpMessage.splitlines():
+		prefix = re.search(r'(?<=§7){}[\w ]*(?=§)'.format(Prefix), line)
+		if prefix is not None:
+			print_message(server, info, RText(line).set_click_event(RAction.suggest_command, prefix.group()), prefix='')
+		else:
+			print_message(server, info, line, prefix='')
 
 
 def on_info(server, info):
-	info2 = copy.deepcopy(info)
-	info2.isPlayer = info2.is_player
-	onServerInfo(server, info2)
+	if not info.is_user:
+		if info.content == 'Saved the game':
+			global game_saved
+			game_saved = True
+		return
+
+	command = info.content.split()
+	if len(command) == 0 or command[0] != Prefix:
+		return
+
+	cmd_len = len(command)
+
+	# MCDR permission check
+	global MinimumPermissionLevel
+	if hasattr(server, 'MCDR') and cmd_len >= 2 and command[0] in MinimumPermissionLevel.keys():
+		if server.get_permission_level(info) < MinimumPermissionLevel[command[0]]:
+			print_message(server, info, '§cPermission denied§r')
+			return
+
+	# !!qb
+	if cmd_len == 1:
+		print_help_message(server, info)
+
+	# !!qb make [<comment>]
+	elif cmd_len >= 2 and command[1] == 'make':
+		comment = info.content.replace('{} make'.format(Prefix), '', 1).lstrip(' ')
+		create_backup(server, info, comment if len(comment) > 0 else None)
+
+	# !!qb back [<slot>]
+	elif cmd_len in [2, 3] and command[1] == 'back':
+		restore_backup(server, info, command[2] if cmd_len == 3 else '1')
+
+	# !!qb confirm
+	elif cmd_len == 2 and command[1] == 'confirm':
+		confirm_restore(server, info)
+
+	# !!qb abort
+	elif cmd_len == 2 and command[1] == 'abort':
+		trigger_abort(server, info)
+
+	# !!qb list
+	elif cmd_len == 2 and command[1] == 'list':
+		list_backup(server, info)
+
+	# !!qb delete
+	elif cmd_len in [2, 3] and command[0] == 'del':
+		delete_backup(server, info, command[2] if cmd_len == 3 else '1')
+
+	else:
+		print_message(server, info, command_run(
+			'Unknown command, input §7{}§r for help'.format(Prefix),
+			'click to check help message',
+			Prefix
+		))
 
 
 def on_load(server, old):
-	server.add_help_message(Prefix, 'back up / restore with {} slots'.format(SlotCount))
+	server.add_help_message(Prefix, command_run('§aback up§r/§crestore§r your world with §6{}§r slots'.format(SlotCount), 'click to check help message', Prefix))
 
 
 def on_unload(server):
